+++ samples/keybinder-unity-examples/Assets/Examples/Example02/ExampleScript02.cs
@@ -6,22 +6,15 @@ public class ExampleScript02 : MonoBehaviour
 {
     [SerializeField] TextMeshProUGUI textComp;
 
-    // Start is called before the first frame update
     void Start()
     {
-        KeyCode[] keysToFilter =
-        {
-            KeyCode.Escape,
-            KeyCode.Numlock,
-        };
-
-        var f = new InputFilter(keysToFilter);
         textComp.text = "Press a key";
         KeyDetector.InputCheckSetActive(true);
+        KeyDetector.KeyReceived += KeyBinderKeyReceived;
     }
 
-    private void Update()
+    private void KeyBinderKeyReceived(KeyCode obj)
     {
-        textComp.text = $"Key pressed: {KeyDetector.LatestKey}";
+        textComp.text = $"Key pressed: {obj}";
     }
 }
