+++ samples/keybinder-unity-examples/Assets/Examples/ExampleDebug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 33cf7cb82435c8545b4499893570e3ea
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
