+++ samples/keybinder-unity-examples/Assets/Examples/ExampleDebug/ExampleDebugScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5532b04ca15d76547ab0647945f68aaa
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
