+++ samples/keybinder-unity-examples/Assets/Examples/ExampleDebug/ExampleDebugScript.cs
@@ -1,0 +1,16 @@
+using KeyBinder;
+using UnityEngine;
+
+public class ExampleDebugScript : MonoBehaviour
+{
+    void Start()
+    {
+        KeyDetector.InputCheckSetActive(true); // turns on input checking
+        KeyDetector.KeyReceived += DebugKey;
+    }
+
+    void DebugKey(KeyCode keyCode)
+    {
+        Debug.Log(KeyDetector.LatestKey); // logs to the console the latest pressed key
+    }
+}
