+++ samples/keybinder-unity-examples/Assets/Examples/ExampleDebug/ExampleDebugScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35cae5801267d8d4aa3b3ad3cb5f5d96
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
